+++ internal/hysteria2/hysteria2.go
@@ -1,11 +1,14 @@
 package hysteria2
 
 import (
+	"errors"
 	"net/url"
 	"strconv"
+
+	"github.com/aceberg/unbox/internal/tls"
 )
 
-// ParseHyst2 converts Hysteria2 URL to struct
+// ParseHyst2 converts Hysteria2 URL string to struct
 func ParseHyst2(raw string) (*Hysteria2, error) {
 	u, err := url.Parse(raw)
 	if err != nil {
@@ -25,11 +28,15 @@ func ParseHyst2(raw string) (*Hysteria2, error) {
 		Server:   u.Hostname(),
 		Port:     portInt,
 		Password: u.User.Username(),
-		TLS: TLS{
-			Enabled:  true,
-			SNI:      q.Get("sni"),
-			Insecure: true,
-		},
+	}
+
+	t, ok := tls.Get(q)
+	if ok {
+		res.TLS = &t
+	}
+
+	if !ok || res.Server == "" || res.Port == 0 {
+		return nil, errors.New("required field empty in " + raw)
 	}
 
 	return res, nil