+++ internal/hysteria2/types.go
@@ -1,18 +1,13 @@
 package hysteria2
 
-// TLS for Hysteria2 config struct
-type TLS struct {
-	Enabled  bool   `json:"enabled"`
-	SNI      string `json:"server_name"`
-	Insecure bool   `json:"insecure"`
-}
+import "github.com/aceberg/unbox/internal/tls"
 
 // Hysteria2 config
 type Hysteria2 struct {
-	Type     string `json:"type"`
-	Tag      string `json:"tag"`
-	Server   string `json:"server"`
-	Port     int    `json:"server_port"`
-	Password string `json:"password"`
-	TLS      TLS    `json:"tls"`
+	Type     string   `json:"type"`
+	Tag      string   `json:"tag"`
+	Server   string   `json:"server"`
+	Port     int      `json:"server_port"`
+	Password string   `json:"password,omitempty"`
+	TLS      *tls.TLS `json:"tls"`
 }