+++ internal/transport/tourl.go
@@ -1,0 +1,35 @@
+package transport
+
+import "net/url"
+
+// ToURL converts Transport struct to url.Values
+func ToURL(tr *Transport) url.Values {
+	q := url.Values{}
+
+	if tr == nil {
+		return q
+	}
+
+	switch tr.Type {
+	case "quic":
+		q.Set("type", "quic")
+
+	case "grpc":
+		q.Set("type", "grpc")
+		if tr.ServName != "" {
+			q.Set("serviceName", tr.ServName)
+		}
+
+	case "ws", "httpupgrade":
+		q.Set("type", tr.Type)
+		if tr.Path != "" {
+			q.Set("path", tr.Path)
+		}
+		if tr.Head != nil && tr.Head.Host != "" {
+			q.Set("host", tr.Head.Host)
+		}
+
+	}
+
+	return q
+}