+++ internal/trojan/types.go
@@ -1,12 +1,9 @@
 package trojan
 
-import "github.com/aceberg/unbox/internal/transport"
-
-// TLS for Trojan config struct
-type TLS struct {
-	Enabled bool   `json:"enabled"`
-	SNI     string `json:"server_name"`
-}
+import (
+	"github.com/aceberg/unbox/internal/tls"
+	"github.com/aceberg/unbox/internal/transport"
+)
 
 // Trojan config
 type Trojan struct {
@@ -15,6 +12,6 @@ type Trojan struct {
 	Server   string               `json:"server"`
 	Port     int                  `json:"server_port"`
 	Password string               `json:"password"`
-	TLS      TLS                  `json:"tls"`
+	TLS      *tls.TLS             `json:"tls,omitempty"`
 	Trans    *transport.Transport `json:"transport,omitempty"`
 }