+++ internal/vless/vless.go
@@ -5,10 +5,11 @@ import (
 	"net/url"
 	"strconv"
 
+	"github.com/aceberg/unbox/internal/tls"
 	"github.com/aceberg/unbox/internal/transport"
 )
 
-// ParseVLESS converts VLESS URL to struct
+// ParseVLESS converts VLESS URL string to struct
 func ParseVLESS(raw string) (*VLESS, error) {
 	u, err := url.Parse(raw)
 	if err != nil {
@@ -22,14 +23,6 @@ func ParseVLESS(raw string) (*VLESS, error) {
 
 	q := u.Query()
 
-	var tls *TLS
-	if q.Get("sni") != "" {
-		tls = &TLS{
-			Enabled: true,
-			SNI:     q.Get("sni"),
-		}
-	}
-
 	res := &VLESS{
 		Type:    "vless",
 		Tag:     u.Fragment,
@@ -37,12 +30,15 @@ func ParseVLESS(raw string) (*VLESS, error) {
 		Port:    portInt,
 		UUID:    u.User.Username(),
 		Flow:    q.Get("flow"),
-		TLS:     tls,
 		PackEnc: "xudp",
 	}
 
+	if res.Server == "" || res.Port == 0 || res.UUID == "" {
+		return nil, errors.New("required field empty in " + raw)
+	}
+
 	if res.Flow != "" && res.Flow != "xtls-rprx-vision" {
-		return nil, errors.New("Unsupported flow: " + res.Flow)
+		return nil, errors.New("unsupported flow: " + res.Flow)
 	}
 
 	tr, ok := transport.Get(q)
@@ -50,26 +46,9 @@ func ParseVLESS(raw string) (*VLESS, error) {
 		res.Trans = &tr
 	}
 
-	sec := q.Get("security")
-	if sec == "reality" {
-		tls := res.TLS
-
-		if tls == nil {
-			tls = &TLS{
-				Enabled: true,
-			}
-		}
-
-		tls.Real = &Reality{
-			Enabled: true,
-			Key:     q.Get("pbk"),
-			ID:      q.Get("sid"),
-		}
-		tls.Utls = &UTLS{
-			Enabled: true,
-			Finger:  q.Get("fp"),
-		}
-		res.TLS = tls
+	t, ok := tls.Get(q)
+	if ok {
+		res.TLS = &t
 	}
 
 	return res, nil